refactor: move getTemplateValues to main process
