+++ src/ambient.d.ts
@@ -1,5 +1,6 @@
 import * as MonacoType from 'monaco-editor';
 
+import { EditorValues } from './interfaces';
 import { App } from './renderer/app';
 
 declare global {
@@ -8,6 +9,7 @@ declare global {
       app: App;
       appPaths: Record<string, string>;
       arch: string;
+      getTemplateValues: (name: string) => Promise<EditorValues>;
       monaco: typeof MonacoType;
       platform: string;
     };
