+++ src/main/templates.ts
@@ -1,7 +1,12 @@
 import * as path from 'path';
 
+// eslint-disable-next-line import/no-unresolved
+import { IpcMainEvent } from 'electron/main';
+
 import { EditorValues } from '../interfaces';
+import { IpcEvents } from '../ipc-events';
 import { readFiddle } from '../utils/read-fiddle';
+import { ipcMainManager } from './ipc';
 
 const STATIC_DIR =
   process.env.NODE_ENV === 'production'
@@ -18,4 +23,11 @@ export function getTemplateValues(name: string): Promise<EditorValues> {
   const templatePath = path.join(STATIC_DIR, 'show-me', name.toLowerCase());
 
   return readFiddle(templatePath);
+}
+
+export function setupTemplates() {
+  ipcMainManager.handle(
+    IpcEvents.GET_TEMPLATE_VALUES,
+    (_: IpcMainEvent, name: string) => getTemplateValues(name),
+  );
 }
