+++ src/preload/preload.ts
@@ -12,6 +12,9 @@ export async function setupFiddleGlobal() {
     app: null as any, // will be set in main.tsx
     appPaths: await ipcRenderer.invoke(IpcEvents.GET_APP_PATHS),
     arch: process.arch,
+    getTemplateValues: (name: string) => {
+      return ipcRenderer.invoke(IpcEvents.GET_TEMPLATE_VALUES, name);
+    },
     monaco: null as any, // will be set in main.tsx
     platform: process.platform,
   };
