+++ tests/mocks/electron-fiddle.ts
@@ -7,6 +7,7 @@ export class ElectronFiddleMock {
     home: `~`,
   };
   public arch = process.arch;
+  public getTemplateValues = jest.fn();
   public monaco = new MonacoMock();
   public platform = process.platform;
 }
